requirements.txt
django==3.0
django-filter
